requirements.txt
tensorflow==1.13.1
keras==2.2.4
pillow==5.1.0
opencv-python==4.1.1.26
matplotlib==2.2.2